fix: passes if we allow v to be fp16 instead of fp8. otherwise we have error < 0.1

+++ flash_attn/flash_attn_triton_amd/fwd_prefill.py
@@ -102,9 +102,11 @@ def _attn_fwd_inner(acc, l_i, m_i, q, k_ptrs, v_ptrs, bias_ptrs, stride_kn, stri
             k_offs_n = None
         k_offs_k = None if not PADDED_HEAD else tl.arange(0, BLOCK_DMODEL)
         k = load_fn(k_ptrs, k_offs_k, k_offs_n, ACTUAL_BLOCK_DMODEL, actual_seqlen_k)
+        # k = k.to(tl.float16)
         if PRE_LOAD_V:
             # We can use the same offsets as k, just with dims transposed.
             v = load_fn(v_ptrs, k_offs_n, k_offs_k, actual_seqlen_k, ACTUAL_BLOCK_DMODEL)
+            v = v.to(tl.float16)
         qk = tl.zeros([BLOCK_M, BLOCK_N], dtype=tl.float32)
         # We start from end of seqlen_k so only the first iteration would need
         # to be checked for padding if it is not a multiple of block_n
@@ -186,11 +188,12 @@ def _attn_fwd_inner(acc, l_i, m_i, q, k_ptrs, v_ptrs, bias_ptrs, stride_kn, stri
         acc = acc * alpha[:, None]
         if not PRE_LOAD_V:
             v = load_fn(v_ptrs, k_offs_n, k_offs_k, actual_seqlen_k, ACTUAL_BLOCK_DMODEL)
+            v = v.to(tl.float16)
         # -- update m_i and l_i
         l_i = l_i * alpha + l_ij
         # update m_i and l_i
         m_i = m_ij
-        acc += tl.dot(p.to(v.type.element_ty), v) #(v.to(tl.float32) * V_SCALE)
+        acc += tl.dot(p.to(v.type.element_ty), v).to(tl.float16) #(v.to(tl.float32) * V_SCALE)
         k_ptrs += BLOCK_N * stride_kn
         v_ptrs += BLOCK_N * stride_vk
         if bias_ptrs is not None:
@@ -418,6 +421,7 @@ def attn_fwd(Q, K, V, bias, Q_SCALE: tl.constexpr, K_SCALE: tl.constexpr, V_SCAL
     if PADDED_HEAD:
         q_ptrs_mask = q_ptrs_mask & (offs_d[None, :] < ACTUAL_BLOCK_DMODEL)
     q = tl.load(q_ptrs, mask=q_ptrs_mask, other=0.0)
+    # q = q.to(tl.float16)
 
     # Here we compute how many full and masked blocks we have.
     padded_block_k = n_extra_tokens != 0
@@ -562,6 +566,8 @@ def attention_prefill_forward_triton_impl(
         v_scale = v.to(torch.float32).abs().max().item()
     else:
         q_scale = k_scale = v_scale = 1
+
+    # force v_scale = 1
     q_scale = k_scale = v_scale = 1
 
     if DEBUG:
